+++ sparklines/lib/chart/sparklines_painter.dart
@@ -1,7 +1,9 @@
 
 import 'package:flutter/material.dart';
+import 'chart_padding.dart';
 import '../data/layout_data.dart';
 import '../interfaces/chart_flip.dart';
+import '../interfaces/chart_padding.dart';
 import '../interfaces/chart_rotation.dart';
 import '../interfaces/chart_transform.dart';
 import '../interfaces/layout.dart';
@@ -12,6 +14,7 @@ class SparklinesPainter extends CustomPainter {
   final List<ISparklinesData> charts;
   final IChartLayout defaultLayout;
   final bool defaultCrop;
+  final ChartPadding padding;
   final double width;
   final double height;
   final List<ISparklinesData>? oldCharts;
@@ -20,6 +23,7 @@ class SparklinesPainter extends CustomPainter {
     required this.charts,
     required this.defaultLayout,
     required this.defaultCrop,
+    this.padding = const ChartPadding(),
     required this.width,
     required this.height,
     this.oldCharts,
@@ -69,9 +73,11 @@ class SparklinesPainter extends CustomPainter {
         canvas.clipRect(bounds);
       }
 
+      final preliminaryMatrix = chartLayout.transform(dimensions);
+      final pathTransform = paddedChartMatrix(preliminaryMatrix, dimensions, padding);
       final transform = ChartTransform(
           dimensions: dimensions,
-          pathTransform: chartLayout.transform(dimensions),
+          pathTransform: pathTransform,
       );
 
       final p0 = transform.xy(0, 0);
@@ -128,6 +134,10 @@ class SparklinesPainter extends CustomPainter {
     }
 
     if (oldDelegate.defaultCrop != defaultCrop) {
+      return true;
+    }
+
+    if (oldDelegate.padding != padding) {
       return true;
     }